fixed compiler errors
git-svn-id: https://ipvs.informatik.uni-stuttgart.de/SGpp/repos/trunk@3321 4eea3252-f0fb-4393-894d-40516dce545b

+++ src/sgpp/parallel/datadriven/basis/common/ocl/OCLKernelImplBase.cpp
@@ -202,7 +202,7 @@ namespace sg {
       kernel_multTrans = new cl_kernel[num_devices];
 
       // initialize arrays
-      for (int i = 0; i < num_devices; i++) {
+      for (size_t i = 0; i < num_devices; i++) {
         command_queue[i] = NULL;
 
         clData[i] = NULL;
